rtm: move some tests around
I think. May have got this wrong.
